refactorisation de FirestationService et des repository

+++ api/src/main/java/com/safetynet/api/model/FireStation.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class FireStation {
 
     private String address;
-    private byte station;
+    private Long station;
 }
